requirements.txt
mock
pygame
pySerial
RPi.GPIO
